requirements.txt
streamlit==1.28.1
pandas==2.1.0
plotly==5.17.0
requests==2.31.0
python-nmap==0.0.1
python-dotenv==1.0.0
fpdf==1.7.2
